requirements.txt
absl-py==0.11.0
astor==0.8.1
backcall==0.2.0
cachetools==4.2.1
certifi==2020.12.5
chardet==4.0.0
cycler==0.10.0
decorator==4.4.2
future==0.18.2
gast==0.2.2
google-auth==1.27.0
google-auth-oauthlib==0.4.2
google-pasta==0.2.0
grpcio==1.35.0
h5py==2.10.0
idna==2.10
importlib-metadata==3.7.0
ipython==7.16.1
ipython-genutils==0.2.0
jedi==0.18.0
joblib==0.16.0
kiwisolver==1.2.0
Markdown==3.3.4
matplotlib==3.3.1
numpy==1.19.5
oauthlib==3.1.0
opencv-python==4.2.0.32
opt-einsum==3.3.0
parso==0.8.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.2.0
pkg-resources==0.0.0
prompt-toolkit==3.0.8
protobuf==3.15.2
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pygments==2.7.3
pyparsing==2.4.7
python-dateutil==2.8.1
PyYAML==5.3.1
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
scikit-learn==0.23.2
scipy==1.4.1
six==1.15.0
tensorboard==2.4.1
tensorboard-plugin-wit==1.8.0
tensorboardX==2.1
termcolor==1.1.0
threadpoolctl==2.1.0
torch==1.6.0
torchvision==0.7.0
tqdm==4.48.2
traitlets==4.3.3
typing-extensions==3.7.4.3
urllib3==1.26.3
wcwidth==0.2.5
Werkzeug==1.0.1
wrapt==1.12.1
zipp==3.4.0
